s02e05/requirements.txt
python-dotenv
requests
langchain
langchain-openai
tiktoken
